+++ openstack/cce/v1/nodes/update.go
@@ -5,24 +5,7 @@ import (
 	"strings"
 
 	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/openstack"
 )
-
-// List returns collection of nodes.
-func List(client *golangsdk.ServiceClient, clusterID string) (r ListResult) {
-	raw, err := client.Get(
-		fmt.Sprintf("https://%s.%s", clusterID, client.ResourceBaseURL()[8:])+strings.Join([]string{"nodes"}, "/"),
-		nil, openstack.StdRequestOpts())
-	return
-}
-
-// Get retrieves a particular nodes based on its unique ID and cluster ID.
-func Get(client *golangsdk.ServiceClient, clusterID, k8sName string) (r GetResult) {
-	raw, err := client.Get(
-		fmt.Sprintf("https://%s.%s", clusterID, client.ResourceBaseURL()[8:])+strings.Join([]string{"nodes", k8sName}, "/"),
-		nil, openstack.StdRequestOpts())
-	return
-}
 
 // UpdateOptsBuilder allows extensions to add additional parameters to the
 // Update request.
